added some design to resolution select box, removed this box from the main page, and made it visible to all of the pages which are using the wallpaperList component, and also refactored the tagpage, its now using the WallpaperList component
