+++ src/components/layout/Dropdown.css
@@ -1,14 +1,17 @@
 .box {
-    margin-inline-start: 43%;
+    text-align: center;
     width:auto;
+    margin: 2%;
+    padding: 1%;
 }
 
 .box select {
+    border: 1px solid black;
+    border-radius: 5px;
     background-color: #aaaaaa;
     color: black;
     padding: 12px;
     width: 250px;
-    border: none;
     font-size: 20px;
     box-shadow: 0 5px 25px rgba(0, 0, 0, 0.2);
     -webkit-appearance: button;
