+++ src/components/pages/MainPage/MainPage.js
@@ -1,47 +1,14 @@
-import React, {useState, useEffect} from "react";
-import {useHttp} from "../../../hooks/http";
-import {Link, BrowserRouter as Router, Route} from "react-router-dom";
-import loadingGif from "../../../loading2.gif";
+import React, {useState} from "react";
+import {Link} from "react-router-dom";
 import WallpaperList from "../WallpaperList";
-import TagPage from "../TagPage/TagPage";
 import './mainPage.css';
 import '../../layout/Dropdown.css';
 
-
-
 const MainPage = () => {
-
-    const [value, setValue] = useState('something');
     const [url, setUrl] = useState('https://wallhaven.cc/api/v1/search?categories=100&purity=100&sorting=random&order=desc');
 
-    const handleSubmit = (event) => {
-        let result = "&resolutions=" + value;
-        setUrl('https://wallhaven.cc/api/v1/search?categories=100&purity=100&sorting=random&order=desc' + result);
-        console.log(url);
-        event.preventDefault();
-    }
-
-    const handleChange = () => {
-        let resolutionValue = document.getElementById("dropdown").value;
-        setValue(resolutionValue);
-    }
-
-    let mainContent = (
+    return (
         <div>
-            <div>
-                <div className="box">
-                    <form onSubmit={handleSubmit}>
-                        <label>
-                            <select id="dropdown" value={value} onChange={handleChange}>
-                                <option value="1920x1080">1920x1080</option>
-                                <option value="1680x1050">1680x1050</option>
-                                <option value="1280x720">1280x720</option>
-                            </select>
-                        </label>
-                        <input type="submit" value="Submit" />
-                    </form>
-                </div>
-            </div>
             <div className={"tag-container"}>
                 <Link to={"/tag/537"}><p>#League of Legends</p></Link>
                 <Link to={"/tag/65348"}><p>#4K</p></Link>
@@ -78,10 +45,7 @@ const MainPage = () => {
             </div>
         </div>
 
-    )
-
-
-    return mainContent;
+    );
 }
 
 export default MainPage;
