+++ src/components/pages/SingleWallpaper/SingleWallpaper.css
@@ -2,7 +2,7 @@
     border: 1px solid black;
     border-radius: 5px;
     padding: 10px 20px;
-    margin: 5px;
+    margin: 1%;
     background-color: #aaaaaa;
 }
